# packages/shared/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
...
  /api/v1/groups:
    get:
      tags:
      - group-controller
      operationId: list_7
      parameters:
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: groupType
        in: query
        required: false
        schema:
          type: string
      - name: poolOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
    post:
      tags:
      - group-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroup"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/groups/{id}/staffing:
    post:
      tags:
      - group-controller
      operationId: staffing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffingRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/groups/{id}/qrcode/rotate:
    post:
      tags:
      - group-controller
      operationId: rotateQrcode
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/groups/{id}:
    get:
      tags:
      - group-controller
      operationId: get_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
    patch:
      tags:
      - group-controller
      operationId: patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGroup"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/groups/{id}/staffing/preview:
    get:
      tags:
      - group-controller
      operationId: preview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: personalWechatId
        in: query
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
...
    ApiResponseMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
...
    CreateGroup:
      required:
      - groupType
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        groupType:
          type: string
        city:
          type: string
        region:
          type: string
        builderAccountId:
          type: string
        targetCapacity:
          type: integer
          format: int32
    StaffingRequest:
      type: object
      properties:
        wecomEmployeeId:
          type: integer
          format: int64
        wecomAccountId:
          type: string
        personalWechatId:
          type: string
...
    PatchGroup:
      type: object
      properties:
        projectId:
          type: string
        builderAccountId:
          type: string
        targetCapacity:
          type: integer
          format: int32
        status:
          type: string
    ApiResponseListMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
...